+++ sdk/storage/azure-storage-file-share/azure/storage/fileshare/aio/_share_client_async.py
@@ -8,7 +8,7 @@ from typing import ( # pylint: disable=unused-import
     Optional, Union, Dict, Any, Iterable, TYPE_CHECKING
 )
 
-from azure.core.exceptions import HttpResponseError
+from azure.core.exceptions import HttpResponseError, ResourceExistsError
 from azure.core.tracing.decorator import distributed_trace
 from azure.core.tracing.decorator_async import distributed_trace_async
 from azure.core.pipeline import AsyncPipeline
@@ -84,7 +84,7 @@ class ShareClient(AsyncStorageAccountHostsMixin, ShareClientBase):
             credential=credential,
             loop=loop,
             **kwargs)
-        self._client = AzureFileStorage(url=self.url, pipeline=self._pipeline, loop=loop)
+        self._client = AzureFileStorage(self.url, base_url=self.url, pipeline=self._pipeline, loop=loop)
         self._client._config.version = get_api_version(kwargs)  # pylint: disable=protected-access
         self._loop = loop
 
@@ -194,7 +194,7 @@ class ShareClient(AsyncStorageAccountHostsMixin, ShareClientBase):
             Only valid for NFS shares. Possible values include: 'NoRootSquash', 'RootSquash', 'AllSquash'.
         :paramtype root_squash: str or ~azure.storage.fileshare.ShareRootSquash
         :returns: Share-updated property dict (Etag and last modified).
-        :rtype: dict(str, Any)
+        :rtype: None
 
         .. admonition:: Example:
 
@@ -231,6 +231,39 @@ class ShareClient(AsyncStorageAccountHostsMixin, ShareClientBase):
                 **kwargs)
         except HttpResponseError as error:
             process_storage_error(error)
+
+    @distributed_trace_async
+    async def create_if_not_exists(self, **kwargs):
+        # type: (Any) -> None
+        """Creates a new Share under the account. If a share with the
+        same name already exists, it is not changed.
+
+        :keyword Dict(str,str) metadata:
+            Name-value pairs associated with the share as metadata.
+        :keyword int quota:
+            The quota to be allotted.
+        :keyword access_tier:
+            Specifies the access tier of the share.
+            Possible values: 'TransactionOptimized', 'Hot', 'Cool'
+        :keyword access_tier: str or ~azure.storage.fileshare.models.ShareAccessTier
+
+            .. versionadded:: 12.4.0
+
+        :keyword int timeout:
+            The timeout parameter is expressed in seconds.
+        :keyword protocols:
+            Protocols to enable on the share. Only one protocol can be enabled on the share.
+        :keyword protocols: str or ~azure.storage.fileshare.ShareProtocols
+        :keyword root_squash:
+            Root squash to set on the share.
+            Only valid for NFS shares. Possible values include: 'NoRootSquash', 'RootSquash', 'AllSquash'.
+        :keyword root_squash: str or ~azure.storage.fileshare.ShareRootSquash
+        :rtype: None
+        """
+        try:
+            return await self.create_share(**kwargs)
+        except ResourceExistsError:
+            return None
 
     @distributed_trace_async
     async def create_snapshot( # type: ignore